Allow a plugin to define multiple Blueprints and relax naming assumption (#164)
This gives more flexibility in naming the plugin, and less worrying about consistently naming plugins and blueprints in different places (thereby facilitating renaming plugins).

The possibility of defining multiple Blueprints in a plugin is more of a perk, but I can imagine it comes in handy if you want to register separate Blueprints for your UI and API, for example, with a different url_prefix.

From PR#164.