Remember the selected time period within the user's session.
I chose to set the session data with the /chart instead of the /chart_data endpoints (for sensors and assets), because the former endpoint can be called without the need to call the latter (using include_data=true). Right now both the sensor and asset page are calling the /chart endpoint with each date pick.