fix: convert mixed offsets to UTC before converting
From pandas.to_datetime docs:

when utc=False (default) and the input is an array-like or Series containing mixed naive/aware datetime, or aware with mixed time offsets. Note that this happens in the (quite frequent) situation when the timezone has a daylight savings policy. In that case you may wish to use utc=True.

Signed-off-by: F.N. Claessen <felix@seita.nl>