fix: Pyomo deprecation
WARNING  pyomo.core.base.param:deprecation.py:238 DEPRECATED: Param 'commitment_quantity' declared with an implicit domain of 'Any'. The default domain for Param objects is 'Any'. However, we will be changing that default to 'Reals' in the future. If you really intend the domain of this Param to be 'Any', you can suppress this warning by explicitly specifying 'within=Any' to the Param constructor.  (deprecated in 5.6.9, will be removed in (or after) 6.0)

Signed-off-by: F.N. Claessen <felix@seita.nl>